requirements.txt
gunicorn==19.9.0
chromedriver-binary==105.0.5195.52
Flask==2.0.3
geckodriver-autoinstaller==0.1.0
ipykernel==6.15.1
ipython==7.34.0
Jinja2==3.1.2
lxml==4.9.1
python-dotenv==0.21.0
selenium==4.4.3
webdriver-manager==3.4.2
chromedriver-autoinstaller
pandas==1.3.5



